chore: reorganize chunk stuff into chunk package
