+++ aspaper-server/src/main/java/com/infernalsuite/asp/level/chunk/FastChunkPruner.java
@@ -1,4 +1,4 @@
-package com.infernalsuite.asp.level;
+package com.infernalsuite.asp.level.chunk;
 
 import ca.spottedleaf.moonrise.patches.chunk_system.level.entity.ChunkEntitySlices;
 import ca.spottedleaf.moonrise.patches.chunk_system.scheduling.NewChunkHolder;
