+++ aspaper-server/src/main/java/com/infernalsuite/asp/level/moonrise/CommonLoadTask.java
@@ -1,6 +1,5 @@
-package com.infernalsuite.asp.level;
+package com.infernalsuite.asp.level.moonrise;
 
-import ca.spottedleaf.concurrentutil.executor.PrioritisedExecutor;
 import ca.spottedleaf.concurrentutil.util.Priority;
 
 public interface CommonLoadTask {
